fixup shaders to have denoise+screen space shadows together
maybe not the ideal experience, this is combining light+shadows and denoising result.
may be better to denoise shadows and then combine with lighting?

using only 1 spatial pass. it is enough to remove noise in this case. (hardly a trying case!)

with 5 or 6 passes of spatial denoising on the lighting result here, the shadows look really garbage.

probably need to improve the depth weighting